backend/requirements.txt
alembic==1.11.1
asyncpg==0.27.0
bcrypt==3.2.2
fastapi==0.95.2
pyjwt==2.8.0
pydantic[email]==1.10.7
passlib==1.7.4
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
SQLAlchemy==1.4.49
uvicorn[standard]==0.22.0